+++ x/leverage/keeper/oracle_test.go
@@ -43,14 +43,15 @@ func (m *mockOracleKeeper) MedianOfHistoricMedians(ctx sdk.Context, denom string
 	return p, uint32(numStamps), nil
 }
 
-func (m *mockOracleKeeper) GetExchangeRate(_ sdk.Context, denom string) (sdk.Dec, error) {
+func (m *mockOracleKeeper) GetExchangeRate(_ sdk.Context, denom string) (oracletypes.ExchangeRate, error) {
 	p, ok := m.symbolExchangeRates[denom]
 	if !ok {
 		// This error matches oracle behavior on missing asset price
-		return sdk.ZeroDec(), oracletypes.ErrUnknownDenom.Wrap(denom)
+		return oracletypes.ExchangeRate{}, oracletypes.ErrUnknownDenom.Wrap(denom)
 	}
 
-	return p, nil
+	// TODO: add timestamp
+	return oracletypes.ExchangeRate{Rate: p}, nil
 }
 
 // Clear clears a denom from the mock oracle, simulating an outage.