+++ CHANGELOG.md
@@ -50,6 +50,10 @@ Ref: https://keepachangelog.com/en/1.0.0/
 
 - [2261](https://github.com/umee-network/umee/pull/2261) Use go 1.21
 
+### Bug Fixes
+
+- [2260](https://github.com/umee-network/umee/pull/2260) fix(oracle): avg params storage.
+
 ## [v6.0.2](https://github.com/umee-network/umee/releases/tag/v6.0.2) - 2023-09-20
 
 ### BugFix